removed core_service_installs.py in favor of the agent_additions file that was already there.
