+++ volttrontesting/platform/test_aip.py
@@ -51,17 +51,3 @@ def aip(request):
     aip.setup()
     return aip
 
-#
-# @pytest.mark.dev
-# def test_can_create(aip):
-#     assert aip
-#     assert os.path.exists(aip.env.volttron_home)
-#
-# @pytest.mark.dev
-# def test_can_install(aip):
-#     pw = PlatformWrapper()
-#     wheel = pw.build_agentpackage(agent_dir="examples/ListenerAgent")
-#     assert wheel
-#     auuid = aip.install_agent(wheel)
-#     assert auuid
-
