+++ volttrontesting/testutils/test_platformwrapper.py
@@ -61,7 +61,7 @@ import pytest
 import time
 import os
 
-from volttron.platform import get_services_core
+from volttron.platform import get_services_core, get_examples
 from volttron.platform.agent import json as jsonapi
 
 from volttrontesting.utils.platformwrapper import start_wrapper_platform, \
@@ -126,7 +126,7 @@ def test_can_install_listener(volttron_instance):
     assert vi is not None
     assert vi.is_running()
 
-    auuid = vi.install_agent(agent_dir="examples/ListenerAgent",
+    auuid = vi.install_agent(agent_dir=get_examples("ListenerAgent"),
                              start=False)
     assert auuid is not None
     started = vi.start_agent(auuid)
@@ -170,7 +170,7 @@ def test_resinstall_agent(volttron_instance):
     for i in range(0,50):
         print("Counter: {}".format(i))
         # auuid = volttron_instance.install_agent(
-        #     agent_dir="examples/ListenerAgent",
+        #     agent_dir=get_examples("ListenerAgent",
         #     vip_identity='test_listener',
         #     start=True)
         auuid = volttron_instance.install_agent(
@@ -234,7 +234,7 @@ def test_can_remove_agent(volttron_instance):
 
     # Install ListenerAgent as the agent to be removed.
     agent_uuid = volttron_instance.install_agent(
-        agent_dir="examples/ListenerAgent", start=False)
+        agent_dir=get_examples("ListenerAgent"), start=False)
     assert agent_uuid is not None
     started = volttron_instance.start_agent(agent_uuid)
     assert started is not None
@@ -312,7 +312,7 @@ def test_can_install_listener_on_two_platforms(get_volttron_instances):
     global messages
     clear_messages()
     auuid = wrapper1.install_agent(
-        agent_dir="examples/ListenerAgent",
+        agent_dir=get_examples("ListenerAgent"),
         start=False)
     assert auuid is not None
     started = wrapper1.start_agent(auuid)
@@ -329,7 +329,7 @@ def test_can_install_listener_on_two_platforms(get_volttron_instances):
 
     clear_messages()
     auuid2 = wrapper2.install_agent(
-        agent_dir="examples/ListenerAgent",
+        agent_dir=get_examples("ListenerAgent"),
         start=True)
     assert auuid2 is not None
     started2 = wrapper2.start_agent(auuid2)
