+++ volttrontesting/testutils/test_platformwrapper2.py
@@ -3,6 +3,7 @@ import pytest
 import time
 import warnings
 
+from volttron.platform import get_examples
 from volttrontesting.fixtures.volttron_platform_fixtures import get_rand_port
 from volttrontesting.utils.platformwrapper import PlatformWrapper
 from volttron.platform.vip.agent import Agent, PubSub, Core
@@ -22,7 +23,7 @@ def test_can_cleanup_installed_listener():
     assert wrapper is not None
     assert wrapper.is_running()
 
-    auuid = wrapper.install_agent(agent_dir="examples/ListenerAgent",
+    auuid = wrapper.install_agent(agent_dir=get_examples("ListenerAgent"),
                                   vip_identity="listener",
                                   start=False)
     assert auuid is not None
